rag/requirements.txt

datasets==4.3.0
faiss-cpu==1.12.0
langchain-text-splitters==0.3.11
numpy==2.3.4
pandas==2.3.3
sentence-transformers==5.1.2
tqdm==4.67.1